qbinplots

This package creates plots using quantile binning.

plot types

qbin_boxplot() creates quantile bin boxplots, highlighting the change in median between qbins.

qbin_barplot() creates quantile bin barplots highlighting the size of means of qbins.

qbin_heatmap() creates quantile bin heatmaps highlighting the distribution within the qbins.

funq_plot() creates a functional view of the data, plotting the median and interquartile range of numerical variables and level frequency of the other variables as a function of the x variable using quantile bins.

Author

Maintainer: Edwin de Jonge edwindjonge@gmail.com (ORCID)